wip: get sql connection and basic queries working

+++ src/domain/entity/claim.rs
@@ -5,7 +5,7 @@ use super::Expiry;
 #[derive(Debug, Serialize, Deserialize)]
 pub struct Claims {
     pub sub: String,
-    pub exp: usize,
+    pub exp: u64,
 }
 
 impl Claims {
