+++ src/domain/entity/expiry.rs
@@ -9,7 +9,7 @@ pub enum Expiry {
 }
 
 impl Expiry {
-    pub fn time(&self) -> usize {
+    pub fn time(&self) -> u64 {
         match self {
             Expiry::Second(count) => Self::expiry_util(*count),
             Expiry::Day(count) => Self::day_expiry(*count),
@@ -29,18 +29,18 @@ impl Expiry {
     }
 
     pub fn is_expired(&self) -> bool {
-        self.time() < Self::now() as usize
+        self.time() < Self::now()
     }
 
     pub fn is_past_epoch_seconds(seconds: usize) -> bool {
         seconds < Self::now() as usize
     }
 
-    fn expiry_failure_fallback() -> usize {
+    fn expiry_failure_fallback() -> u64 {
         10000000000
     }
 
-    fn expiry_util(seconds: u64) -> usize {
+    fn expiry_util(seconds: u64) -> u64 {
         let timestamp = Self::now() + seconds;
 
         match timestamp.try_into() {
@@ -52,22 +52,22 @@ impl Expiry {
         }
     }
 
-    fn day_expiry(count: u64) -> usize {
+    fn day_expiry(count: u64) -> u64 {
         let day_in_sec = 86_400;
         Self::expiry_util(day_in_sec * count)
     }
 
-    fn week_expiry(count: u64) -> usize {
+    fn week_expiry(count: u64) -> u64 {
         let week_in_sec = 604_800;
         Self::expiry_util(week_in_sec * count)
     }
 
-    fn month_expiry(count: u64) -> usize {
+    fn month_expiry(count: u64) -> u64 {
         let month_in_sec = 2_678_400;
         Self::expiry_util(month_in_sec * count)
     }
 
-    fn year_expiry(count: u64) -> usize {
+    fn year_expiry(count: u64) -> u64 {
         let year_in_sec = 31_536_000;
         Self::expiry_util(year_in_sec * count)
     }
